Consolidate Operand code
